wyoming_deepgram_stt/requirements.txt
wyoming==1.5.3
deepgram-sdk==3.2.5